Improved performance for batch add & update
Modification of the add(List<Object> objects) and update(List<Object> objects) functions.
Instead of calling the add(Object object) which was opening and closing the database on each object, we are now opening the database and beginning the transaction once, and closing it after all the objects were inserted. 
I tried the change on my data set. Before It was taking me more than 1 minute to add all the data set in the database. After the modification it takes me now not more than 1 second.